# src/ApiGateways/AdminPanel/AdminPanel.Infrastructure/Specifications/ServiceForumSwagger.json
{
...
  "swagger": "2.0",
  "info": {
    "title": "County RP Forum Service API",
    "description": "The County RP Forum Service API documentation description.",
    "version": "v1"
  },
  "host": "localhost:10511",
  "schemes": [
    "https"
  ],
  "paths": {
    "/Forum": {
      "post": {
        "tags": [
          "Forum"
        ],
        "operationId": "Forum_Create",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "apiForumDtoIn",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ApiForumDtoIn"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "201": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ApiForumDtoOut"
            }
          },
          "400": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      },
      "get": {
        "tags": [
          "Forum"
        ],
        "operationId": "Forum_Get",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ApiForumDtoOut"
              }
            }
          },
          "404": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      }
    },
    "/Forum/Hierarchical": {
      "get": {
        "tags": [
          "Forum"
        ],
        "operationId": "Forum_GetHierarchical",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ApiHierarchicalForumDtoOut"
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      }
    },
    "/Forum/{id}": {
      "get": {
        "tags": [
          "Forum"
        ],
        "operationId": "Forum_GetById",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ApiForumDtoOut"
            }
          },
          "404": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      },
      "put": {
        "tags": [
          "Forum"
        ],
        "operationId": "Forum_Edit",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "apiForumDtoIn",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ApiForumDtoIn"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          "404": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Forum"
        ],
        "operationId": "Forum_Delete",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      }
    },
...
    "/Forum/Ordered": {
      "put": {
        "tags": [
          "Forum"
        ],
        "operationId": "Forum_EditOrdered",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "apiUpdatedOrderedForumsDtoIn",
            "in": "body",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ApiUpdatedOrderedForumDtoIn"
              }
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      }
    },
...
  },
  "definitions": {
    "ApiForumDtoOut": {
      "type": "object",
      "required": [
        "id",
        "parentId",
        "order"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        },
        "parentId": {
          "type": "integer",
          "format": "int32"
        },
        "order": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "ApiForumDtoIn": {
      "type": "object",
      "required": [
        "parentId",
        "order"
      ],
      "properties": {
        "name": {
          "type": "string"
        },
        "parentId": {
          "type": "integer",
          "format": "int32"
        },
        "order": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "ApiHierarchicalForumDtoOut": {
      "type": "object",
      "required": [
        "id",
        "parentId",
        "order"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        },
        "parentId": {
          "type": "integer",
          "format": "int32"
        },
        "order": {
          "type": "integer",
          "format": "int32"
        },
        "childForums": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ApiHierarchicalForumDtoOut"
          }
        }
      }
    },
...
    "ApiUpdatedOrderedForumDtoIn": {
      "type": "object",
      "required": [
        "id",
        "parentId",
        "order"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "parentId": {
          "type": "integer",
          "format": "int32"
        },
        "order": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "apiKey": {
      "type": "apiKey",
      "description": "Copy 'Bearer ' + valid api key into field",
      "name": "Authorization",
      "in": "header"
    }
  }
}